requirements.txt
mkdocs-material==9.5.29
mkdocs-material[imaging]==9.5.29
mkdocs-awesome-pages-plugin==2.9.2
mkdocs-redirects==1.2.1
mkdocs-video==1.5.0
mkdocs-macros-plugin==1.0.5
python-frontmatter==1.1.0
